+++ service/controller/catalog/resource/helmrepository/resource.go
@@ -1,0 +1,43 @@
+package helmrepository
+
+import (
+	"github.com/giantswarm/microerror"
+	"github.com/giantswarm/micrologger"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+const (
+	name = "helmrepository"
+)
+
+type empty struct{}
+
+type Config struct {
+	CtrlClient client.Client
+	Logger     micrologger.Logger
+}
+
+type Resource struct {
+	ctrlClient client.Client
+	logger     micrologger.Logger
+}
+
+func (r *Resource) Name() string {
+	return name
+}
+
+func New(config Config) (*Resource, error) {
+	if config.CtrlClient == client.Client(nil) {
+		return nil, microerror.Maskf(invalidConfigError, "%T.CtrlClient must not be empty", config)
+	}
+	if config.Logger == nil {
+		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
+	}
+
+	r := &Resource{
+		ctrlClient: config.CtrlClient,
+		logger:     config.Logger,
+	}
+
+	return r, nil
+}